split the reducer by features
