+++ stock-tracker/src/features/keyStats/component/keyStats.js
@@ -4,7 +4,7 @@ import Loading from "../../loading/component/loading";
 import "./KeyStats.css";
 
 const KeyStats = () => {
-  const response = useSelector(state => state.response);
+  const response = useSelector(state => state.keyStats.response);
   const loading = useSelector(state => state.loading);
   const error = useSelector(state => state.error.stockData);
