+++ stock-tracker/src/features/keyStats/redux/actionTypes.js
@@ -1,0 +1,1 @@
+export const ADD_RESPONSE = "ADD_RESPONSE";
