+++ stock-tracker/src/features/keyStats/redux/actions.js
@@ -1,0 +1,6 @@
+import { ADD_RESPONSE } from "./actionTypes";
+
+export const addResponseAction = response => ({
+    type: ADD_RESPONSE,
+    payload: response
+  });
