+++ stock-tracker/src/features/keyStats/redux/keyStatsReducer.js
@@ -1,0 +1,20 @@
+import {
+    ADD_RESPONSE
+  } from "./actionTypes";
+  
+  const initialState = {
+    response: false,
+  };
+  
+  export const keyStatsReducer = (state = initialState, action) => {
+    switch (action.type) {
+      case ADD_RESPONSE:
+        return {
+          ...state,
+          response: action.payload,
+        };
+      default:
+        return state;
+    }
+  }
+
