+++ stock-tracker/src/features/latestNews/redux/actionTypes.js
@@ -1,0 +1,1 @@
+export const ADD_NEWS = "ADD_NEWS";
