+++ stock-tracker/src/features/latestNews/redux/actions.js
@@ -1,0 +1,6 @@
+import { ADD_NEWS } from "./actionTypes"
+
+export const addLatestNewsAction = latestNews => ({
+    type: ADD_NEWS,
+    payload: latestNews
+  });
