+++ stock-tracker/src/features/overview/redux/actionTypes.js
@@ -1,0 +1,1 @@
+export const ADD_COMPANY_OVERVIEW = "ADD_COMPANY_OVERVIEW";
