+++ stock-tracker/src/features/overview/redux/actions.js
@@ -1,0 +1,6 @@
+import { ADD_COMPANY_OVERVIEW } from "./actionTypes"
+
+export const addCompanyOverviewAction = companyOverview => ({
+    type: ADD_COMPANY_OVERVIEW,
+    payload: companyOverview
+  });
