+++ stock-tracker/src/features/search/redux/actionTypes.js
@@ -1,0 +1,3 @@
+export const ADD_SYMBOL = "ADD_SYMBOL";
+export const ADD_SEARCH_INPUT = "ADD_SEARCH_INPUT";
+export const ADD_SUGGESTIONS = "ADD_SUGGESTIONS";
