+++ stock-tracker/src/features/topPeers/redux/actionTypes.js
@@ -1,0 +1,1 @@
+export const ADD_TOP_PEERS = "ADD_TOP_PEERS";
