+++ stock-tracker/src/features/topPeers/redux/actions.js
@@ -1,0 +1,9 @@
+import { ADD_TOP_PEERS } from "./actionTypes"
+  
+  export const addTopPeersAction = topPeers => ({
+    type: ADD_TOP_PEERS,
+    payload: topPeers
+  });
+  
+
+
