+++ stock-tracker/src/features/topPeers/redux/peersReducer.js
@@ -1,0 +1,18 @@
+import { ADD_TOP_PEERS } from "./actionTypes";
+  
+  const initialState = {
+    topPeers: []
+  };
+  
+  export const peersReducer = (state = initialState, action) => {
+    switch (action.type) {
+      case ADD_TOP_PEERS:
+        return {
+          ...state,
+          topPeers: action.payload
+        };
+      default:
+        return state;
+    }
+  }
+
