+++ stock-tracker/src/store/actionTypes.js
@@ -1,9 +1,2 @@
-export const ADD_RESPONSE = "ADD_RESPONSE";
-export const ADD_SYMBOL = "ADD_SYMBOL";
 export const ADD_COMPANIES = "ADD_COMPANIES";
-export const ADD_COMPANY_OVERVIEW = "ADD_COMPANY_OVERVIEW";
-export const ADD_NEWS = "ADD_NEWS";
-export const ADD_TOP_PEERS = "ADD_TOP_PEERS";
-export const ADD_SEARCH_INPUT = "ADD_SEARCH_INPUT";
-export const ADD_SUGGESTIONS = "ADD_SUGGESTIONS";
 export const INITIAL_STARTUP = "INITIAL_STARTUP";
