+++ stock-tracker/src/store/rootReducer.js
@@ -1,7 +1,21 @@
 import { combineReducers } from "redux";
 import { chartReducer } from "../features/chart/redux/chartReducer";
-import { errorReducer } from "../features/error/redux/errorReducer"
+import { errorReducer } from "../features/error/redux/errorReducer";
+import { keyStatsReducer } from "../features/keyStats/redux/keyStatsReducer";
+import { newsReducer } from "../features/latestNews/redux/newsReducer";
+import { overviewReducer } from "../features/overview/redux/overviewReducer";
+import { searchReducer } from "../features/search/redux/searchReducer";
+import { peersReducer } from "../features/topPeers/redux/peersReducer";
 
 import { reducer } from "./reducer";
 
-export const rootReducer = combineReducers({reducer: reducer, chart: chartReducer, error: errorReducer})
+export const rootReducer = combineReducers({
+    reducer: reducer, 
+    chart: chartReducer, 
+    error: errorReducer, 
+    keyStats: keyStatsReducer,
+    news: newsReducer,
+    overview: overviewReducer,
+    search: searchReducer,
+    peers: peersReducer
+})
